refactoring build process

+++ bin/build.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
 import argparse, glob, os, platform, shutil, subprocess, sys
+import xml.etree.ElementTree as ET
 
 
 def safe_delete( dp ):
@@ -20,6 +21,8 @@ parser = argparse.ArgumentParser( description = "Build a redistributable package
 parser.add_argument( "-d", "--dir", required = True )
 parser.add_argument( "-e", "--exeprefix", required = True )
 parser.add_argument( "-p", "--platform", required = True, choices = [ "windows", "android" ] )
+parser.add_argument( "--debug", action = "store_true", default = False )
+parser.add_argument( "--install", action = "store_true", default = False )
 args = parser.parse_args()
 
 DEVDIR = os.path.dirname( os.path.realpath( __file__ ) )
@@ -69,6 +72,67 @@ def build_windows():
 	# resources
 	shutil.copytree( DEVP( RESOURCE_PATH ), TGTP( RESOURCE_PATH ) )
 
+class AndroidPaths:
+	def which( self, path ):
+		# python bugs, yay
+		if os.name == "nt":
+			for ext in os.environ.get( "PATHEXT", "" ).split( os.pathsep ):
+				if os.path.exists( path + ext.lower() ):
+					return path + ext.lower()
+		return shutil.which( path, path = os.pathsep + os.environ[ "PATH" ] )
+	def envpath( self, printname, name ):
+		p = os.environ.get( name )
+		if p is None:
+			print( "  ERROR: Missing environment variable %s" % name )
+			self.errors = True
+		elif p == "" or not os.path.exists( p ):
+			print( "  ERROR: Path specified in %s env.var. does not exist" % name )
+			self.errors = True
+		else:
+			print( "- %s: %s" % ( printname, p ) )
+		return p
+		
+	def exepath( self, printname, printbase, *pathbits ):
+		testpath = os.path.join( *pathbits )
+		p = self.which( testpath )
+		if p is None:
+			print( "  ERROR: Could not find '%s' in '%s' (tried %s)" % ( printname, printbase, testpath ) )
+			self.errors = True
+		else:
+			print( "- %s: %s" % ( printname, p ) )
+		return p
+		
+	def __init__( self ):
+		self.errors = False
+		
+		# TODO
+		os.environ[ "JAVA_HOME" ] = "C:/Program Files/Java/jdk1.8.0_101"
+		
+		print( "Checking Android build environment..." )
+		asn = "Android SDK"
+		self.sdk = self.envpath( asn, "ANDROID_SDK" )
+		self.adb = self.exepath( "adb", asn, self.sdk, "platform-tools", "adb" )
+		self.android = self.exepath( "android", asn, self.sdk, "tools", "android" )
+		# TODO find build tools
+		self.zipalign = self.exepath( "zipalign", asn, self.sdk, "build-tools", "25.0.2", "zipalign" )
+		
+		jhn = "Java"
+		self.javahome = self.envpath( jhn, "JAVA_HOME" )
+		self.keytool = self.exepath( "keytool", jhn, self.javahome, "bin", "keytool" )
+		self.jarsigner = self.exepath( "jarsigner", jhn, self.javahome, "bin", "jarsigner" )
+		
+		self.ant = self.which( "ant" )
+		if self.ant is None:
+			print( "ant not in path" )
+			self.errors = True
+		else:
+			print( "- ant: " + self.ant )
+		
+		if self.errors:
+			raise Exception( "Failed to validate Android build environment" )
+		else:
+			print( "Done" )
+
 def android_ignore_prebuilt( directory, contents ):
 	if DEVP( directory ) == DEVP( "android" ):
 		return [ "bin", "gen", "jni", "obj" ]
@@ -82,42 +146,46 @@ def build_android():
 	# scripts
 	copy_scripts( TGTP( "android_tmp/assets" ) )
 	
-	# TODO
-	android_SDK_location = "E:/android/sdk"
-	ant_cmd_path = "E:/android/ant/bin/ant.bat"
-	os.environ[ "JAVA_HOME" ] = "C:/Program Files/Java/jdk1.8.0_101"
-	android_cmd_path = os.path.join( android_SDK_location, "tools", "android.bat" )
-	keytool_path = os.path.join( os.environ[ "JAVA_HOME" ], "bin", "keytool.exe" )
-	jarsigner_path = os.path.join( os.environ[ "JAVA_HOME" ], "bin", "jarsigner.exe" )
-	zipalign_path = os.path.join( android_SDK_location, "build-tools", "25.0.2", "zipalign.exe" )
+	paths = AndroidPaths()
 	# build APK
 	cwd = TGTP( "android_tmp" )
-	run_process([ android_cmd_path, "update", "project", "--path", ".", "--target", "android-16" ], cwd )
-	run_process([ ant_cmd_path, "release" ], cwd )
-	# copy built APK to output path
-	apkname = TGTP( "%s-release.apk" % args.exeprefix )
-	shutil.copy( TGTP( "android_tmp/bin/sgsouzel-release-unsigned.apk" ), apkname )
-	# sign APK
-	run_process([ keytool_path,
-		"-genkey", "-v",
-		"-keystore", "my-release-key.keystore",
-		"-alias", "alias_name",
-		"-keyalg", "RSA",
-		"-keysize", "2048",
-		"-validity", "10000"
-	], cwd )
-	run_process([ jarsigner_path,
-		"-verbose",
-		"-sigalg", "SHA1withRSA",
-		"-digestalg", "SHA1",
-		"-keystore", "my-release-key.keystore",
-		apkname,
-		"alias_name"
-	], cwd )
-	run_process([ jarsigner_path, "-verify", "-verbose", "-certs", apkname ])
-	run_process([ zipalign_path, "-v", "4", apkname, apkname + ".aligned.zip" ])
-	os.remove( apkname )
-	os.rename( apkname + ".aligned.zip", apkname )
+	run_process([ paths.android, "update", "project", "--path", ".", "--target", "android-16" ], cwd )
+	if args.debug:
+		run_process([ paths.ant, "debug" ], cwd )
+		# copy built APK to output path
+		apkname = TGTP( "%s-debug.apk" % args.exeprefix )
+		shutil.copy( TGTP( "android_tmp/bin/sgsouzel-debug.apk" ), apkname )
+	else:
+		run_process([ paths.ant, "release" ], cwd )
+		# copy built APK to output path
+		apkname = TGTP( "%s-release.apk" % args.exeprefix )
+		shutil.copy( TGTP( "android_tmp/bin/sgsouzel-release-unsigned.apk" ), apkname )
+		# sign APK
+		# TODO allow generating key and passing own key separately
+		run_process([ paths.keytool,
+			"-genkey", "-v",
+			"-keystore", "my-release-key.keystore",
+			"-alias", "alias_name",
+			"-keyalg", "RSA",
+			"-keysize", "2048",
+			"-validity", "10000"
+		], cwd )
+		run_process([ paths.jarsigner,
+			"-verbose",
+			"-sigalg", "SHA1withRSA",
+			"-digestalg", "SHA1",
+			"-keystore", "my-release-key.keystore",
+			apkname,
+			"alias_name"
+		], cwd )
+		run_process([ paths.jarsigner, "-verify", "-verbose", "-certs", apkname ])
+		run_process([ paths.zipalign, "-v", "4", apkname, apkname + ".aligned.zip" ])
+		os.remove( apkname )
+		os.rename( apkname + ".aligned.zip", apkname )
+	if args.install:
+		# manifest = ET.parse( DEVP( "android", "AndroidManifest.xml" ) ).getroot()
+		# run_process([ paths.adb, "uninstall", manifest.attrib[ "package" ] ])
+		run_process([ paths.adb, "install", "-r", apkname ])
 
 
 if args.platform == "windows":
